+++ CHANGELOG.md
@@ -6,6 +6,14 @@ x.y.z Release notes (yyyy-MM-dd)
 <!-- ### Fixes - Include, if needed -->
 <!-- ### Internal - Include, if needed -->
 
+### Enhancements
+
+* Implement the `isSignedIn` PWA communication handler which returns `true` by default if the authorization is valid in `IDKit`. This behavior can be overridden in `AppCallback.isSignedIn(isSignedIn)`.
+
+### Fixes
+
+* Regenerate communication code to fix problem with unparsable error response
+
 
 10.0.0 Release notes (2021-09-16)
 =============================================================